Implemented a basic autonomous movement system
Added hummingbird function to calculate speed on a bell curve for smoother movement.
Created an AutoDriver class to manage autonomous movement.
Refactored the Auto and Owlin opmodes to use the new system. Moved Auto to extend LinearOpMode due to stop() issue.
Disabled telemetry output for odometry in Owlin and Diagnostics.